requirements.txt
Django==1.6.2
Markdown==2.4
MySQL-python==1.2.5
Pillow==2.3.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-filebrowser==3.5.4
django-filter==0.7
django-grappelli==2.5.1
django-imagekit==3.2
django-tinymce==1.5.2
django-toolbelt==0.0.1
djangorestframework==2.3.13
gunicorn==19.1.1
pilkit==1.1.6
psycopg2==2.5.4
requests==2.3.0
six==1.5.2
static3==0.5.1
wsgiref==0.1.2
